refactor: modularize client scripts
